+++ src/media/color/mod.rs
@@ -7,6 +7,7 @@
 
 mod base; // trait ColorBase
 mod namespace; // struct Color
+mod rgb; // Rgb[a][8|16|F32|F64], Rgb[a]Lin[F32|F64]
 
 crate::items! { // structural access: _mods, _all, _always
     #[allow(unused)]
@@ -15,10 +16,9 @@ crate::items! { // structural access: _mods, _all, _always
     pub use _always::*;
 
     mod _mods {
-        pub use super::{base::*, namespace::*};
+        pub use super::{base::*, namespace::*, rgb::_all::*};
         // WIPZONE
         // pub use super::gray::*;
-        // pub use super::rgb::_all::*;
         // pub use super::spectral::*;
     }
     pub(super) mod _all {
@@ -30,5 +30,4 @@ crate::items! { // structural access: _mods, _all, _always
 }
 // WIPZONE
 // mod gray;
-// mod rgb;
 // mod spectral;