+++ DOCS/CHANGELOG.md
@@ -43,6 +43,10 @@
 ## media
 - delete: `[Audio|Color|Draw|Font|Media|Midi][Result|Error]`.
 
+### color
+- add types: `Rgb[a][8|16|F32|F64]`, `Rgb[a]Lin[F32|F64]`
+- add module: `media::color::rgb`.
+
 ## num
 ### float
 - add: